Separate deploy and package logic from deploy command
